+++ apps/decodex/src/orchestrator/status_run_projection.rs
@@ -1,48 +1,59 @@
 //! Operator run status projection, protocol/activity readback, and lane lifecycle metrics.
 
-use super::{
-	ARCHITECTURE_RECOVERY_PACKET_EVENT_TYPE, ARCHITECTURE_RECOVERY_STARTED_EVENT_TYPE,
-	ARCHITECTURE_RECOVERY_TERMINAL_EVENT_TYPE, AUTHORITY_BOUNDARY_CHECK_EVENT_TYPE,
-	AUTHORITY_DECISION_REQUEST_EVENT_TYPE, AgentPrivateEvidenceRef,
-	CONTINUATION_PENDING_RUN_STATUS, ChildAgentActivityBucket, ChildAgentActivitySummary,
-	CodexAccountActivitySummary, Duration, EXECUTION_LIVENESS_PROCESS_IDENTITY_MISMATCH, HashMap,
-	HashSet, OperatorArchitectureRecoveryStatus, OperatorAuthorityDecisionRequestStatus,
-	OperatorBoundaryStatus, OperatorContinuationRecoveryStatus, OperatorHistoryLaneStatus,
-	OperatorLaneControlProjection, OperatorLaneLifecycleAttemptEvidence,
-	OperatorLaneLifecycleMetrics, OperatorLaneLifecyclePhaseMetrics, OperatorLifecycleMetricPhase,
-	OperatorLoopStatus, OperatorPhaseAcceptanceStatus, OperatorRecoveryBudgetStatus,
-	OperatorReviewCheckpointStatus, OperatorReviewCheckpointSummaryFields,
-	OperatorReviewLoopStatus, OperatorReviewRouteCount, OperatorRunAppServerState,
-	OperatorRunControlCapability, OperatorRunLifecycleProjection, OperatorRunProtocolSummary,
-	OperatorRunStatus, OperatorRunTiming, OperatorTerminalFinalizeProjection,
-	PHASE_ACCEPTANCE_CHECK_EVENT_TYPE, PHASE_GOAL_RECOVERY_AUTOMATIC_CONTINUATION_LIMIT,
-	PHASE_GOAL_RECOVERY_BLOCKED_EVENT_TYPE, PHASE_GOAL_RECOVERY_EVENT_TYPE, PrivateExecutionEvent,
-	ProjectLoopEvidenceSnapshot, ProjectRunStatus, ProtocolActivitySummary,
-	REVIEW_POLICY_CONVERGENCE_BUDGET, RUN_LEASE_IDLE_TIMEOUT, RUN_OPERATION_AGENT_RUN,
-	RUN_OPERATION_IDLE, RUN_OPERATION_REVIEW_WRITEBACK, RUN_OPERATION_WAITING_EXTERNAL,
-	ReviewLevel, Rfc3339, RunActivityMarker, ServiceConfig, StateStore,
-	TERMINAL_GUARDED_RUN_STATUS, Value, append_primary_account_if_missing,
-	marker_process_liveness_for_marker, not_loaded_history_ledger_outcome, observed_idle_duration,
-	operator_authority_decision_request_status_from_event, operator_autonomy_lineage_statuses,
-	operator_autonomy_objective_status, operator_autonomy_proposal_statuses,
-	operator_autonomy_report_status, operator_autonomy_signal_statuses,
-	operator_run_counts_as_attention, operator_run_counts_as_running,
-	operator_run_has_fresh_execution, operator_run_has_recent_app_server_execution,
-	operator_run_needs_attention, private_evidence_ref_for_run_fields, public_text,
-	relative_worktree_path_for_path, run_activity_idle_timeout, state,
-};
-use time::OffsetDateTime;
-
 mod history;
 mod loop_status;
 mod run;
 mod runtime;
 
-#[allow(clippy::wildcard_imports)]
-pub(super) use history::*;
-#[allow(clippy::wildcard_imports)]
-pub(super) use loop_status::*;
-#[allow(clippy::wildcard_imports)]
-pub(super) use run::*;
-#[allow(clippy::wildcard_imports)]
-pub(super) use runtime::*;
+#[allow(unused_imports)]
+pub(super) use self::{
+	history::{
+		current_lane_lifecycle_attempts, hydrate_current_lane_lifecycle_metrics,
+		hydrate_history_lane_from_run, issue_identifier_from_run_id, issue_identifier_in_text,
+		operator_history_lanes, operator_lane_lifecycle_attempt_evidence,
+		operator_lane_lifecycle_metrics, operator_lane_lifecycle_phase_metrics,
+		operator_lane_lifecycle_totals, operator_lifecycle_metric_phase, operator_run_group_key,
+		operator_run_issue_identifier_from_fields, operator_run_issue_key,
+		operator_run_lifecycle_metric_phase, operator_run_current_lane_snapshot_attempt,
+	},
+	loop_status::{
+		operator_architecture_recovery_next_action, operator_architecture_recovery_status_for_reason,
+		operator_architecture_recovery_status_from_event, operator_boundary_policy_blocks_landing,
+		operator_boundary_policy_decision_from_disposition,
+		operator_boundary_policy_requires_enhanced_evidence, operator_boundary_status_from_event,
+		operator_latest_review_checkpoint_event_status, operator_loop_autonomy,
+		operator_loop_status_for_run, operator_loop_status_for_run_with_evidence,
+		operator_loop_status_next_action, operator_loop_status_summary,
+		operator_review_checkpoint_summary_fields, operator_review_loop_status,
+	},
+	run::{
+		hydrate_operator_run_derived_status, operator_run_accounts,
+		operator_run_active_goal_phase, operator_run_default_review_phase,
+		operator_run_lifecycle_loop_summary, operator_run_lifecycle_projection,
+		operator_run_loop_status, operator_run_phase_acceptance_status,
+		operator_run_private_evidence, operator_run_public_progress_phase,
+		operator_run_relative_worktree_path, operator_run_status, operator_run_status_from_parts,
+		operator_run_wait_reason,
+	},
+	runtime::{
+		classify_operator_run_operation, classify_operator_run_phase,
+		contains_protocol_activity_host_path_shape, contains_protocol_activity_secret_shape,
+		format_optional_i64, format_optional_unix_timestamp, idle_duration_seconds,
+		is_high_entropy_protocol_activity_token, load_operator_run_marker,
+		marker_protocol_summary_supersedes_run, max_optional_i64, operator_continuation_recovery_event_status,
+		operator_latest_repo_gate_failure_progress_diagnostic, operator_protocol_activity_detail_is_public,
+		operator_protocol_event_counts_as_live_execution, operator_repo_gate_failure_progress_diagnostic,
+		operator_run_app_server_state, operator_run_child_agent_activity,
+		operator_run_continuation_recovery_status, operator_run_control_capability,
+		operator_run_execution_liveness, operator_run_has_app_server_execution_evidence,
+		operator_run_has_recent_protocol_execution_evidence, operator_run_is_suspected_stall,
+		operator_run_live_evidence_source, operator_run_progress_diagnostic, operator_run_protocol_activity,
+		operator_run_protocol_summary, operator_run_queue_lease_state,
+		operator_run_status_projection_reason, operator_run_terminal_finalize_projection,
+		operator_run_timing, operator_run_visible_status, process_liveness_reason_is_identity_mismatch,
+		protocol_activity_is_non_work_only, protocol_activity_token_separator,
+		protocol_wait_reason_from_child_bucket, review_handoff_terminal_finalize_wait_reason,
+		review_repair_terminal_finalize_wait_reason, sanitize_operator_protocol_activity_summary,
+		suspected_operator_run_stall_threshold, visible_operator_run_retry_schedule,
+	},
+};
