+++ apps/decodex/src/orchestrator/status_run_projection/runtime.rs
@@ -1,5 +1,3 @@
-#[allow(clippy::wildcard_imports)] use super::*;
-
 mod activity;
 mod continuation;
 mod core;
@@ -8,8 +6,42 @@ mod formatting;
 mod liveness;
 mod terminal_finalize;
 
-#[allow(clippy::wildcard_imports)]
+#[allow(unused_imports)]
 pub(in crate::orchestrator) use self::{
-	activity::*, continuation::*, core::*, diagnostics::*, formatting::*, liveness::*,
-	terminal_finalize::*,
+	activity::{
+		contains_protocol_activity_host_path_shape, contains_protocol_activity_secret_shape,
+		is_high_entropy_protocol_activity_token, operator_protocol_activity_detail_is_public,
+		operator_run_child_agent_activity, operator_run_protocol_activity,
+		protocol_activity_token_separator, sanitize_operator_protocol_activity_summary,
+	},
+	continuation::{
+		operator_continuation_recovery_event_status, operator_run_continuation_recovery_status,
+	},
+	core::{
+		load_operator_run_marker, marker_protocol_summary_supersedes_run,
+		operator_run_app_server_state, operator_run_control_capability,
+		operator_run_protocol_summary, operator_run_timing,
+	},
+	diagnostics::{
+		classify_operator_run_operation, classify_operator_run_phase,
+		operator_latest_repo_gate_failure_progress_diagnostic,
+		operator_repo_gate_failure_progress_diagnostic, operator_run_is_suspected_stall,
+		operator_run_progress_diagnostic, protocol_activity_is_non_work_only,
+		suspected_operator_run_stall_threshold, visible_operator_run_retry_schedule,
+	},
+	formatting::{
+		format_optional_i64, format_optional_unix_timestamp, idle_duration_seconds,
+		max_optional_i64, protocol_wait_reason_from_child_bucket,
+	},
+	liveness::{
+		operator_protocol_event_counts_as_live_execution, operator_run_execution_liveness,
+		operator_run_has_app_server_execution_evidence,
+		operator_run_has_recent_protocol_execution_evidence, operator_run_live_evidence_source,
+		operator_run_queue_lease_state, operator_run_status_projection_reason,
+		operator_run_visible_status, process_liveness_reason_is_identity_mismatch,
+	},
+	terminal_finalize::{
+		operator_run_terminal_finalize_projection, review_handoff_terminal_finalize_wait_reason,
+		review_repair_terminal_finalize_wait_reason,
+	},
 };
