+++ apps/decodex/src/orchestrator/status_run_projection/runtime/continuation.rs
@@ -1,4 +1,8 @@
-#[allow(clippy::wildcard_imports)] use super::*;
+use crate::orchestrator::{
+	OperatorContinuationRecoveryStatus, PHASE_GOAL_RECOVERY_AUTOMATIC_CONTINUATION_LIMIT,
+	PHASE_GOAL_RECOVERY_BLOCKED_EVENT_TYPE, PHASE_GOAL_RECOVERY_EVENT_TYPE, PrivateExecutionEvent,
+	ProjectLoopEvidenceSnapshot, ProjectRunStatus, Value,
+};
 
 pub(in crate::orchestrator) fn operator_run_continuation_recovery_status(
 	loop_evidence: &ProjectLoopEvidenceSnapshot,
