+++ apps/decodex/src/orchestrator/status_run_projection/runtime/formatting.rs
@@ -1,4 +1,4 @@
-#[allow(clippy::wildcard_imports)] use super::*;
+use time::{OffsetDateTime, format_description::well_known::Rfc3339};
 
 pub(in crate::orchestrator) fn protocol_wait_reason_from_child_bucket(
 	current_bucket: &str,
