+++ apps/decodex/src/orchestrator/status_run_projection/runtime/terminal_finalize.rs
@@ -1,4 +1,7 @@
-#[allow(clippy::wildcard_imports)] use super::*;
+use crate::orchestrator::{
+	OperatorTerminalFinalizeProjection, PrivateExecutionEvent, ProjectLoopEvidenceSnapshot,
+	ProjectRunStatus, RUN_OPERATION_REVIEW_WRITEBACK, Value,
+};
 
 pub(in crate::orchestrator) fn operator_run_terminal_finalize_projection(
 	loop_evidence: &ProjectLoopEvidenceSnapshot,
